The page-geometry fix missed the path every real scan takes
A security review of the public repository found that yesterday's overflow fix guards the
wrong half of the decoder, and reading the code confirms it.

PageHeader::validate() bounds what a peer may declare on the wire. That covers fromRaw(),
because the declared numbers only become a page's geometry on the uncompressed path.
Nothing produced in this repository ever sends RawBgr - the encoder emits Jpeg or Png - and
on those paths decode() takes width and height from the codec's own metadata instead:

    frame->GetSize(&w, &h);
    page.width  = static_cast<int32_t>(w);
    page.height = static_cast<int32_t>(h);

A JPEG SOF marker or a PNG IHDR, neither of which the wire format inspects and neither of
which validate() ever sees. So the bound existed, was correct, and was not reached by the
only path anybody uses. decode.h had not been touched since 11 August; the fix went into
rs_protocol.h on the 21st.

DecodedPage::stride() then reproduced the original bug exactly - ((width * bitsPerPixel +
31) / 32) * 4 in int32_t - fed from that unvalidated source. copyBottomUp checked the
resulting size afterwards, which is the wrong order: the check ran on a number the overflow
had already made meaningless, so a width chosen to wrap the product small passed it, and
CopyPixels then wrote the real frame's rows into a buffer sized for the wrapped one. Inside
Acrobat, or an ERP, in the user's own session.

Three changes, and the third is the one that matters:

  - stride() computes in int64_t, so it cannot overflow for any width the type accepts and
    the checks are checking a number that means what it says.
  - DecodedPage::validate() bounds the geometry itself, to the same limits as
    PageHeader::validate().
  - It is called where the geometry is set, not where one of its two sources is parsed -
    after GetSize on the encoded path, and in fromRaw and copyBottomUp as well, so no
    caller can reach the copy without having passed it.

The lesson is the one this project keeps relearning. The check was placed where the data
was easiest to reach rather than where the value was decided, and everything downstream
was assumed covered by it. That assumption held for the path that was tested and not for
the path that ships.

+++ native/ScanBridge.TwainDS/ds.cpp
@@ -215,10 +215,11 @@ void writeUncompressedTiff(Put&& put, const DecodedPage& page) {
 
     // DIB rows are bottom-up and padded; TIFF wants top-down and tight. 24bpp DIBs are BGR
     // while TIFF RGB is R,G,B, so colour samples are swapped as the row is repacked.
-    const int32_t dibStride = page.stride();
+    const int64_t dibStride = page.stride();
     std::vector<uint8_t> row(tiffStride);
     for (int32_t y = 0; y < page.height; ++y) {
-        const uint8_t* source = page.bits.data() + static_cast<size_t>(page.height - 1 - y) * dibStride;
+        const uint8_t* source = page.bits.data()
+                              + static_cast<size_t>(page.height - 1 - y) * static_cast<size_t>(dibStride);
         if (samples == 3) {
             for (int32_t x = 0; x < page.width; ++x) {
                 row[static_cast<size_t>(x) * 3 + 0] = source[static_cast<size_t>(x) * 3 + 2];
